+++ lib/src/generator.dart
@@ -76,6 +76,17 @@ class _NetworkGenerator {
   }
 
   Future<void> _generateNewCaKey() async {
+    final parsedRenew = parseDuration(network.renew);
+    if (parsedRenew != null) {
+      final currentCerts =
+          await loadCertificatesFromDirectory(p.join(outputPath, 'ca', 'keys'));
+      final threshold = DateTime.now().subtract(parsedRenew);
+      if (currentCerts
+          .any((c) => c.certificate.details!.notBefore!.isAfter(threshold))) {
+        return;
+      }
+    }
+
     final tempCAPrefix = p.join(_tempRoot.path, 'ca', _netName);
     await File('$tempCAPrefix.crt').parent.create(recursive: true);
     final newCaCert = await _cli.ca(